knn/requirements.txt
Flask==3.0.2
matplotlib==3.8.3
numpy==1.24.2
pymongo==4.6.2
python-dotenv==1.0.1
scikit_learn==1.4.1.post1
scipy==1.12.0
